NMT_MODULE/requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
asttokens #@ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall #@ file:///home/ktietz/src/ci/backcall_1611930011877/work
blinker==1.6.2
brotlipy==0.7.0
build==0.10.0
CacheControl==0.12.14
cachetools==5.3.1
certifi #@ file:///croot/certifi_1690232220950/work/certifi
cffi #@ file:///work/ci_py311/cffi_1676822533496/work
charset-normalizer #@ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cleo==2.0.1
click==8.1.3
colorama==0.4.6
ConfigArgParse==1.7
contourpy #@ file:///work/ci_py311/contourpy_1676827066340/work
crashtest==0.4.1
cryptography #@ file:///croot/cryptography_1689373676338/work
ctranslate2==3.18.0
cycler #@ file:///tmp/build/80754af9/cycler_1637851556182/work
datasets==2.13.1
debugpy #@ file:///croot/debugpy_1690905042057/work
decorator #@ file:///opt/conda/conda-bld/decorator_1643638310831/work
dill==0.3.6
distlib==0.3.6
dulwich==0.21.5
entmax==1.1
evaluate==0.4.0
executing #@ file:///opt/conda/conda-bld/executing_1646925071911/work
fasttext==0.9.2
fasttext-wheel==0.9.2
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
gast==0.4.0
gmpy2 #@ file:///work/ci_py311/gmpy2_1676839849213/work
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
html5lib==1.1
huggingface-hub==0.16.4
idna #@ file:///work/ci_py311/idna_1676822698822/work
installer==0.7.0
ipykernel #@ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1655369107642/work
ipython #@ file:///croot/ipython_1691532092695/work
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi #@ file:///work/ci_py311_2/jedi_1679336495545/work
jeepney==0.8.0
Jinja2 #@ file:///work/ci_py311/jinja2_1676823587943/work
joblib==1.3.1
jsonargparse==3.13.1
jsonschema==4.17.3
jupyter_client #@ file:///croot/jupyter_client_1680171862562/work
jupyter_core #@ file:///croot/jupyter_core_1679906564508/work
keras==2.13.1
keyring==23.13.1
kiwisolver #@ file:///work/ci_py311/kiwisolver_1676827230232/work
libclang==16.0.6
lightning-utilities==0.9.0
lockfile==0.12.2
lxml==4.9.2
Markdown==3.4.4
MarkupSafe #@ file:///work/ci_py311/markupsafe_1676823507015/work
matplotlib #@ file:///croot/matplotlib-suite_1679593461707/work
matplotlib-inline #@ file:///work/ci_py311/matplotlib-inline_1676823841154/work
mkl-fft==1.3.6
mkl-random #@ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
more-itertools==9.1.0
mpmath #@ file:///croot/mpmath_1690848262763/work
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
munkres==1.1.4
nest-asyncio #@ file:///work/ci_py311/nest-asyncio_1676823382924/work
networkx #@ file:///croot/networkx_1690561992265/work
nltk==3.8.1
numpy #@ file:///croot/numpy_and_numpy_base_1691091611330/work
oauthlib==3.2.2
openai==0.27.8
OpenNMT-py==3.3
opt-einsum==3.3.0
packaging #@ file:///croot/packaging_1678965309396/work
pandas==1.5.3
parso #@ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect #@ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare #@ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==3.8.0
ply==3.11
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
portalocker==2.7.0
prompt-toolkit #@ file:///work/ci_py311/prompt-toolkit_1676824857154/work
# protobuf==3.20.0
psutil==5.9.5
ptyprocess #@ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval #@ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyahocorasick==2.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycparser #@ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyDeprecate==0.3.2
Pygments #@ file:///croot/pygments_1684279966437/work
pyonmttok==1.37.1
pyOpenSSL #@ file:///croot/pyopenssl_1690223430423/work
pyparsing #@ file:///work/ci_py311/pyparsing_1677811559502/work
pyproject_hooks==1.0.0
PyQt5-sip==12.11.0
pyrsistent==0.19.3
PySocks #@ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil #@ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-lightning==1.9.5
pytz==2023.3
PyYAML==6.0
pyzmq #@ file:///croot/pyzmq_1686601365461/work
rapidfuzz==2.15.1
regex==2023.6.3
requests #@ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
rsa==4.9
sacrebleu==2.3.1
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
SecretStorage==3.3.3
sentencepiece==0.1.99
seqeval==1.2.2
shellingham==1.5.0.post1
sip #@ file:///work/ci_py311/sip_1676825117084/work
six #@ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post7
stack-data #@ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy #@ file:///work/ci_py311_2/sympy_1679339311852/work
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-cpu==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
threadpoolctl==3.2.0
thumt==1.2.0
tokenizers==0.13.3
toml #@ file:///tmp/build/80754af9/toml_1616166611790/work
tomlkit==0.11.8
#torch==2.0.1
#torchaudio==2.0.2
torchmetrics==0.8.2
#torchvision==0.15.2
tornado #@ file:///croot/tornado_1690848263220/work
tqdm==4.65.0
traitlets #@ file:///work/ci_py311/traitlets_1676823305040/work
transformers==4.29.0
triton==2.0.0
trove-classifiers==2023.5.24
typing_extensions #@ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
unbabel-comet==2.0.2
urllib3 #@ file:///croot/urllib3_1686163155763/work
virtualenv==20.23.1
waitress==2.1.2
wcwidth #@ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.15.0
xxhash==3.2.0
yarl==1.9.2
